fix minor warning

+++ src/gauri/imgui/imgui_layer.cpp
@@ -143,7 +143,8 @@ void ImGuiLayer::OnUpdate()
 {
     ImGuiIO &io = ImGui::GetIO();
     Application &app = Application::Get();
-    io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+    io.DisplaySize =
+        ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 
     float time = (float)glfwGetTime();
     io.DeltaTime = m_Time > 0.0f ? (time - m_Time) : (1.0f / 60.0f);
